Add files via upload
This was my initial design, with the gaps fitting into the caging on the bag of the fence. But shortly after I came up with a better design which is also posted. (Note none of these are to scale or to fit any phones, these will be optimized after all considered designs are tried.)