Add files via upload
This is my second version of the clamp. The change I made with this was that instead of securing the cage of the fence this hangs above the desired filming point, making it more suitable for more types of back fencing. (This is also not to scale, all measurements will be finalized after considering all designs.)